pyproject.toml
[project]
name = "pyqt"
version = "0.1.0"
description = ""
authors = [
    {name = "JiaLong0209",email = "smt02091395@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11, <=3.13"
dependencies = [
    "pyqt5 (>=5.15.11,<6.0.0)",
    "pyserial (>=3.5,<4.0)",
    "pyautogui (>=0.9.54,<0.10.0)",
    "numpy (>=2.3.5,<3.0.0)",
    "pyqtgraph (>=0.14.0,<0.15.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "joblib (>=1.5.3,<2.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "xgboost (>=3.1.2,<4.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "torch (>=2.9.1,<3.0.0)",
    "torchvision (>=0.24.1,<0.25.0)",
    "tqdm (>=4.67.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
